improvements & bug fixes
